requirement.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
django-filter==23.1
dnspython==2.3.0
mongoengine==0.27.0
psycopg2-binary==2.9.6
pymongo==3.12.3
redis
faker
tzdata

